Move operator logic from endpoints to resource package
