+++ pkg/operator/endpoints/deploy.go
@@ -17,19 +17,11 @@ limitations under the License.
 package endpoints
 
 import (
-	"fmt"
 	"net/http"
 
-	"github.com/cortexlabs/cortex/pkg/consts"
 	"github.com/cortexlabs/cortex/pkg/lib/errors"
 	"github.com/cortexlabs/cortex/pkg/lib/files"
-	"github.com/cortexlabs/cortex/pkg/lib/hash"
-	"github.com/cortexlabs/cortex/pkg/lib/zip"
-	"github.com/cortexlabs/cortex/pkg/operator/config"
 	"github.com/cortexlabs/cortex/pkg/operator/resources"
-	"github.com/cortexlabs/cortex/pkg/operator/schema"
-	"github.com/cortexlabs/cortex/pkg/types"
-	"github.com/cortexlabs/cortex/pkg/types/spec"
 )
 
 func Deploy(w http.ResponseWriter, r *http.Request) {
@@ -55,55 +47,12 @@ func Deploy(w http.ResponseWriter, r *http.Request) {
 		respondError(w, r, err)
 		return
 	}
-	projectID := hash.Bytes(projectBytes)
-	projectKey := spec.ProjectKey(projectID)
-	projectFileMap, err := zip.UnzipMemToMem(projectBytes)
+
+	response, err := resources.DeployAPI(projectBytes, configPath, configBytes, force)
 	if err != nil {
 		respondError(w, r, err)
 		return
 	}
 
-	projectFiles := resources.ProjectFiles{
-		ProjectByteMap: projectFileMap,
-		ConfigFilePath: configPath,
-	}
-	apiConfigs, err := spec.ExtractAPIConfigs(configBytes, types.AWSProviderType, projectFiles, configPath)
-	if err != nil {
-		respondError(w, r, err)
-		return
-	}
-
-	err = resources.ValidateClusterAPIs(apiConfigs, projectFiles)
-	if err != nil {
-		err = errors.Append(err, fmt.Sprintf("\n\napi configuration schema can be found here: https://docs.cortex.dev/v/%s/deployments/api-configuration", consts.CortexVersionMinor))
-		respondError(w, r, err)
-		return
-	}
-
-	isProjectUploaded, err := config.AWS.IsS3File(config.Cluster.Bucket, projectKey)
-	if err != nil {
-		respondError(w, r, err)
-		return
-	}
-	if !isProjectUploaded {
-		if err = config.AWS.UploadBytesToS3(projectBytes, config.Cluster.Bucket, projectKey); err != nil {
-			respondError(w, r, err)
-			return
-		}
-	}
-
-	results := make([]schema.DeployResult, len(apiConfigs))
-	for i, apiConfig := range apiConfigs {
-		api, msg, err := resources.UpdateAPI(&apiConfig, projectID, force)
-		results[i].Message = msg
-		if err != nil {
-			results[i].Error = errors.Message(err)
-		} else {
-			results[i].API = *api
-		}
-	}
-
-	respond(w, schema.DeployResponse{
-		Results: results,
-	})
+	respond(w, response)
 }
